Trait recon_mcts::BaseGD

pub trait BaseGD {
    type Player;
    type State;
    type Action;
    type Score;
    type ActionIter: IntoIterator<Item = (Self::Player, Self::Action)>;
    fn available_actions(
        &self, 
        player: &Self::Player, 
        state: &Self::State
    ) -> Option<Self::ActionIter>;

    fn apply_action(
        &self, 
        state: Self::State, 
        action: &Self::Action
    ) -> Option<Self::State>;

    fn score_leaf(
        &self, 
        parent_score: Option<&Self::Score>, 
        parent_player: &Self::Player, 
        state: &Self::State
    ) -> Option<Self::Score>;
}

A trait that can be used to implemented DynGD without implementing GameDynamics. BaseGD is automatically implemented for types that implemented GameDynamics.

DynGD is a supertrait of BaseGD.

For descriptions of the associated types and required methods see GameDynamics.

Associated Types

type Player

type State

type Action

type Score

type ActionIter: IntoIterator<Item = (Self::Player, Self::Action)>

Loading content...
Required methods

fn available_actions(
&self,
player: &Self::Player,
state: &Self::State
) -> Option<Self::ActionIter>

fn apply_action(
&self,
state: Self::State,
action: &Self::Action
) -> Option<Self::State>

fn score_leaf(
&self,
parent_score: Option<&Self::Score>,
parent_player: &Self::Player,
state: &Self::State
) -> Option<Self::Score>

Loading content...