Trait recon_mcts::DynGD

pub trait DynGD: BaseGD {
    fn select_node(
        &self, 
        parent_score: Option<&Self::Score>, 
        parent_player: &Self::Player, 
        parent_node_state: &Self::State, 
        purpose: SelectNodeState, 
        scores_and_actions: &mut dyn Iterator<Item = (Ref<Option<Self::Score>>, Ref<Self::Action>)>
    ) -> Self::Action;

    fn backprop_scores(
        &self, 
        player: &Self::Player, 
        score_current: Option<&Self::Score>, 
        child_scores: &mut dyn Iterator<Item = Ref<Self::Score>>
    ) -> Option<Self::Score>;
}

A supertrait of BaseGD. Its purpose is to implement GameDynamics for trait objects. Implementing DynGD is only required if dynamic dispatch is needed.

Since GameDynamics::select_node and GameDynamics::backprop_scores have generic parameters, a type implementing GameDynamics cannot be used as a trait object directly. However, GameDynamics is automatically implemented for types that dereference to a type implementing DynGD (see GameDynamics Implementors).

BaseGD is automatically implemented for types that implement GameDynamics, but can also be implemented directly without implementing GameDynamics if static dispatch is not required (i.e. implementing BaseGD and DynGD without implementing GameDynamics will allow for dynamic dispatch but not static dispatch). For descriptions of the associated types and required methods see GameDynamics.

See the provided implementation of Nim for a demonstration of using static and dynamic dispatch.

Implementation Note

Using DynGD for dynamic dispatch does not result in additional heap allocations. This is accomplished via the use of a single RefCell on the stack which is referenced by all Ref<T>s returned by an IntoIterator type in the DynGD trait methods. As a result, calling Deref::deref on a Ref that is not the Ref most recently returned by the iterator will result in a panic. If out-of-sequence access is required, the implementation must store the dereferenced values rather than the Refs.

Required methods

fn select_node(
&self,
parent_score: Option<&Self::Score>,
parent_player: &Self::Player,
parent_node_state: &Self::State,
purpose: SelectNodeState,
scores_and_actions: &mut dyn Iterator<Item = (Ref<Option<Self::Score>>, Ref<Self::Action>)>
) -> Self::Action

fn backprop_scores(
&self,
player: &Self::Player,
score_current: Option<&Self::Score>,
child_scores: &mut dyn Iterator<Item = Ref<Self::Score>>
) -> Option<Self::Score>

Loading content...