Trait recon_mcts::GameDynamics

pub trait GameDynamics {
    type Player;
    type State;
    type Action;
    type Score;
    type ActionIter: IntoIterator<Item = (Self::Player, Self::Action)>;
    fn available_actions(
        &self, 
        player: &Self::Player, 
        state: &Self::State
    ) -> Option<Self::ActionIter>;

    fn apply_action(
        &self, 
        state: Self::State, 
        action: &Self::Action
    ) -> Option<Self::State>;

    fn select_node<II, Q, A>(
        &self, 
        parent_score: Option<&Self::Score>, 
        parent_player: &Self::Player, 
        parent_node_state: &Self::State, 
        purpose: SelectNodeState, 
        scores_and_actions: II
    ) -> Self::Action
    where
        Self: Sized,
        II: Clone + IntoIterator<Item = (Q, A)>,
        Q: Deref<Target = Option<Self::Score>>,
        A: Deref<Target = Self::Action>;

    fn backprop_scores<II, Q>(
        &self, 
        player: &Self::Player, 
        score_current: Option<&Self::Score>, 
        child_scores: II
    ) -> Option<Self::Score>
    where
        Self: Sized,
        II: Clone + IntoIterator<Item = Q>,
        Q: Deref<Target = Self::Score>;

    fn score_leaf(
        &self, 
        parent_score: Option<&Self::Score>, 
        parent_player: &Self::Player, 
        state: &Self::State
    ) -> Option<Self::Score>;
}

Requires implementation by the user in order to provide the rules of the game and desired specifics of the algorithm.

Implementation Note

The associated types should generally not be types that allow interior mutability as mutating them in the methods provided by the GameDynamics trait may result in logic error. However, in some implementations it may be desirable to use a GameDynamics::Score type with interior mutability, e.g. in order to modify the number of times it was selected in GameDynamics::select_node.

Associated Types

type Player

Most likely an enum to designate which player is allowed to select an action at the current node. For a one player game, this type can simply be the unit ().

type State

A type that complete describes the state of the game. The implementation of the GameDynamics must ensure that graph formed by the search tree is acyclic. If the game (e.g. chess) allows the same board position to be reached multiple times, some value within the state could track how many times the state has been reached. This is often required anyway because repeat board position are actually different game states (e.g. in chess, reaching the same board position three times results in a tie, so the game state is actually distinct despite the position of pieces on the board being identical).

type Action

Most likely an enum or integer value to describe the action that transitions from one state to another.

type Score

A type that represents an evaluation of Self::State. This would commonly be a float, but more elaborate types may be useful for some games, e.g. those with more than two players.

type ActionIter: IntoIterator<Item = (Self::Player, Self::Action)>

An iterator used to establish the available actions and next player (note that the player can be different for different actions).

If the returned iterator contains a closure, it will need to be boxed on stable Rust until #![feature(type_alias_impl_trait)] is stabilized (see #63063 and #63066).

Loading content...
Required methods

fn available_actions(
&self,
player: &Self::Player,
state: &Self::State
) -> Option<Self::ActionIter>

Convert an input state into the available actions / moves. Return None if the game is over. If the game is over, the GameDynamics::score_leaf method will be called to evaluate the state

fn apply_action(
&self,
state: Self::State,
action: &Self::Action
) -> Option<Self::State>

Modify the state input (i.e. game board) with an action.

In some games it may be easier to determine whether an action is valid by trying to apply it and then assess whether it succeeds or fails, we allow GameDynamics::apply_action to return on Option::None in case of failure (in such a case the action will be treated as though it was never included in the output of available_actions).

Note that the user must ensure that for any State the same state can never be reached again by applying any number of Actions obtained via available_actions (i.e. the user must ensure the tree graph is acyclic)

fn select_node<II, Q, A>(
&self,
parent_score: Option<&Self::Score>,
parent_player: &Self::Player,
parent_node_state: &Self::State,
purpose: SelectNodeState,
scores_and_actions: II
) -> Self::Action where
Self: Sized,
II: Clone + IntoIterator<Item = (Q, A)>,
Q: Deref<Target = Option<Self::Score>>,
A: Deref<Target = Self::Action>,

Select the action to take based on the scores.

Note that because scores_and_actions are received in non-deterministic order, the result of the MCTS algorithm may also be non-deterministic if there is a tie with respect to the selected score for two different actions.

The SelectNodeState maybe be used to add an additional exploration parameter when searching for the best action.

parent_score is included because MCTS commonly uses the parent's count of number of visits in the selection phase.

parent_node_state is not typically a required argument, but can be helpful if performing only a partial expansion of nodes is desired upon reaching a leaf node and recalculating the state is preferred over storing it in an enum. See additional info in GameDynamics::score_leaf.

fn backprop_scores<II, Q>(
&self,
player: &Self::Player,
score_current: Option<&Self::Score>,
child_scores: II
) -> Option<Self::Score> where
Self: Sized,
II: Clone + IntoIterator<Item = Q>,
Q: Deref<Target = Self::Score>,

Score a parent node based on its child nodes.

score_current is passed to enable comparison with the calculated score so that if the difference is insignificant, the backpropagation to the root node can be cut off by returning a None. For deep trees this may provide a decent performance boost. The root note does not have a score initially, so an Option<&Self::Score> is used for score_current.

Returning a None does not imply that score_current will be None on the next iteration -- that would only be the case iff it was a None on the current iteration and a None is returned.

fn score_leaf(
&self,
parent_score: Option<&Self::Score>,
parent_player: &Self::Player,
state: &Self::State
) -> Option<Self::Score>

Take a leaf node's state and assign the node a score, whether via simulation or otherwise.

Note that for a leaf where no actions are possible (i.e. a terminal node) the score should be considered immutable. Changes to the score of such a terminal node on subsequent simulation runs will not result in the updates being backpropagated through the tree.

Implementation Note:

The library currently expands all branches of a leaf, though this can effectively be modified to a delayed scoring / rollout by using a type such as the following for the associated type Score and then evaluating the Score in GameDynamics::select_node if a MyScore::State is encountered.

// pseudo-code
impl GameDynamics for MyGame {
    // `MyRealScore` will need to be updated via a shared ref (i.e. it must be either an
    // atomic variable or wrapped in a type that provides interior mutability). This is an
    // exception to the rule that the associated types should not be mutated by the trait
    // methods for cases where a delayed rollout is desired.
    type Score = enum MyScore {
        State(Self::State),
        Score(MyRealScore),
    };
    ...
}

Alternatively, one could return a None and derive the state in GameDynamics::select_node using GameDynamics::available_actions if a None is encountered. The state will be calculated for the leaf node irrespective of whether the scoring is performed by GameDynamics::score_leaf because it is used in checking whether the same state already exists in the tree (i.e. encountered via a different sequence of actions).

Loading content...