Trait recon_mcts::OnDrop

pub trait OnDrop {
    fn on_drop(self_arc: ArcWrap<Self>);
}

A trait used to remove nodes from the transposition table that are no longer reachable from the root. Generally for internal use.

Required methods

fn on_drop(self_arc: ArcWrap<Self>)

Loading content...