Trait recon_mcts::StateMemory

pub trait StateMemory {
    type State;
    fn eq(&self, rhs: &Self) -> bool;

    fn modify_state(state: &RwLock<Option<Self::State>>);
}

A trait used to modify how states are stored in the transposition table. Generally for internal use.

Associated Types

type State

The State of the Node as specified in GameDynamics::State.

Loading content...
Required methods

fn eq(&self, rhs: &Self) -> bool

Specifies the implementation of PartialEq for a Node with a particular StateMemory mixin.

fn modify_state(state: &RwLock<Option<Self::State>>)

Associated function to set how the State should be stored after a child is created.

Loading content...