Trait recon_mcts::BaseGD

pub trait BaseGD {
    type Player;
    type State;
    type Action;
    type Score;
    type ActionIter: IntoIterator<Item = (Self::Player, Self::Action)>;
    fn available_actions(
        &self, 
        player: &Self::Player, 
        state: &Self::State
    ) -> Option<Self::ActionIter>;

    fn apply_action(
        &self, 
        state: Self::State, 
        action: &Self::Action
    ) -> Option<Self::State>;

    fn score_leaf(
        &self, 
        parent_score: Option<&Self::Score>, 
        parent_player: &Self::Player, 
        state: &Self::State
    ) -> Option<Self::Score>;
}

A trait that can be used to implemented DynGD without implementing GameDynamics. BaseGD is automatically implemented for types that implemented GameDynamics.

DynGD is a supertrait of BaseGD.

For descriptions of the associated types and required methods see GameDynamics.

Associated Types

type Player

See GameDynamics::Player for a description of this associated type.

type State

See GameDynamics::State for a description of this associated type.

type Action

See GameDynamics::Action for a description of this associated type.

type Score

See GameDynamics::Score for a description of this associated type.

type ActionIter: IntoIterator<Item = (Self::Player, Self::Action)>

See GameDynamics::ActionIter for a description of this associated type.

Loading content...
Required methods

fn available_actions(
&self,
player: &Self::Player,
state: &Self::State
) -> Option<Self::ActionIter>

See GameDynamics::available_actions for a description of this associated function.

fn apply_action(
&self,
state: Self::State,
action: &Self::Action
) -> Option<Self::State>

See GameDynamics::apply_action for a description of this associated function.

fn score_leaf(
&self,
parent_score: Option<&Self::Score>,
parent_player: &Self::Player,
state: &Self::State
) -> Option<Self::Score>

See GameDynamics::score_leaf for a description of this associated function.

Loading content...